Add new file src/l10n/l10n-utils.ts and move everything that's not the 'locales' object or related types to this new file. Fix all the imports and types that broke due to this change
