+++ src/components/Perils/Perils.stories.tsx
@@ -1,7 +1,8 @@
 import { select, withKnobs } from '@storybook/addon-knobs'
 import React from 'react'
 import { globalStoryMock } from 'utils/storybook'
-import { locales, fallbackLocale } from 'src/l10n/locales'
+import { locales } from 'src/l10n/locales'
+import { fallbackLocale } from 'src/l10n/l10n-utils'
 import { TypeOfContract } from './types'
 import { Perils } from './'
